pyproject.toml
[project]
name = "walue_whatsapp_provider"
version = "0.0.1"
description = "WhatsApp Calling & Messaging Provider Platform for Frappe/ERPNext"
authors = [
    {name = "Walue Biz", email = "support@walue.biz"}
]
license = {text = "MIT"}
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
    "requests>=2.28.0",
    "pyjwt>=2.6.0",
    "cryptography>=38.0.0"
]

[build-system]
requires = ["flit_core>=3.4"]
build-backend = "flit_core.buildapi"

[tool.bench.frappe-dependencies]
frappe = ">=15.0.0"
